catalog-service/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
alembic
python-dotenv
psycopg2-binary
pytest
httpx
ruff
pytest-cov